requirements.txt
beautifulsoup4==4.12.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.1
idna==3.4
packaging==23.1
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.9.0
Pillow==10.0.0
pycparser==2.21
pytesseract==0.3.10
requests==2.31.0
soupsieve==2.4.1
tqdm==4.65.0
urllib3==2.0.3
Wand==0.6.11
